+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 ---------
 
-#### Next Release
+#### v0.12.6 (2019-12-11)
 - Properly handles case when translation downloading is cancelled
   - As a result, reverted "Only emit if book names and book short names are loaded"
 - Throw exception when missing translation list